+++ lib/widgets/home_widgets/catalog_list.dart
@@ -3,6 +3,7 @@ import 'package:flutter_demo/models/cart.dart';
 import 'package:velocity_x/velocity_x.dart';
 import 'package:flutter_demo/pages/home_detail_page.dart';
 import '../../models/catalog.dart';
+import '../../pages/home_widgets/add_to_cart.dart';
 import '../../widgets/themes.dart';
 // import '../home_page.dart';
 import 'catalog_image.dart';
@@ -65,7 +66,7 @@ class CatalogItem extends StatelessWidget {
                 buttonPadding: EdgeInsets.zero,
                 children: [
                   "\$${catalog.price}".text.bold.xl.make(),
-                  _AddToCart(catalog: catalog)
+                  AddToCart(catalog: catalog)
                 ],
               ).pOnly(right: 8.0)
             ],
@@ -76,38 +77,3 @@ class CatalogItem extends StatelessWidget {
   }
 }
 
-class _AddToCart extends StatefulWidget {
-  final Item catalog;
-
-  const _AddToCart({
-    Key? key,
-    required this.catalog,
-  }) : super(key: key);
-
-  @override
-  State<_AddToCart> createState() => _AddToCartState();
-}
-
-class _AddToCartState extends State<_AddToCart> {
-  bool isAdded = false;
-
-  @override
-  Widget build(BuildContext context) {
-    return ElevatedButton(
-      onPressed: () {
-        isAdded = isAdded.toggle();
-        final _catalog = CatalogModel();
-        final _cart = CartModel();
-        _cart.catalog = _catalog;
-        _cart.add(widget.catalog);
-        setState(() {});
-      },
-      style: ButtonStyle(
-          backgroundColor: MaterialStateProperty.all(context.theme.buttonColor),
-          shape: MaterialStateProperty.all(
-            StadiumBorder(),
-          )),
-      child: isAdded ? Icon(Icons.done) : "Add to Cart".text.make(),
-    );
-  }
-}
